+++ Krowi_AchievementFilter/Localization/enUS.lua
@@ -16,6 +16,7 @@ L["Not installed"] = "Not installed";
 L["T_TAB_TEXT"] = "Achievement Filter";
 
 -- Categories
+L["Current Zone"] = "Current Zone";
 L["Zones"] = "Zones";
 L["Class Hall"] = "Class Hall";
 
@@ -25,6 +26,7 @@ L["Not Obtainable"] = "Not Obtainable";
 L["Covenant"] = "Covenant";
 L["Select all"] = "Select all";
 L["Reverse Sort"] = "Reverse Sort";
+L["Collapse Series"] = "Collapse Series";
 
 -- Options
 L["O_REQUIRES_RELOAD"] = AF_COLOR_ORANGE .. "Requires a reload to take full effect." .. AF_COLOR_END;
@@ -106,9 +108,16 @@ L["FT_FILTER_DESC"] = "The default filtering of achievements is quite basic with
                         AF_TAB .. "- {obtainable}: Show obtainable achievements\n" ..
                         AF_TAB .. "- {notObtainable}: Show not obtainable achievements\n" ..
                         AF_TAB .. "- {faction}\n" ..
-                        AF_TAB .. AF_TAB .. "- {neutral}: Show achievements that both Alliance and Horde can earn\n" ..
-                        AF_TAB .. AF_TAB .. "- {alliance}: Show achievements only Alliance can earn\n" ..
-                        AF_TAB .. AF_TAB .. "- {horde}: Show achievements only Horde can earn\n\n" ..
+                        AF_TAB .. AF_TAB .. "- {neutral}: Show achievements that both {alliance_wc} and {horde_wc} can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {alliance}: Show achievements only {alliance_wc} can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {horde}: Show achievements only {horde_wc} can earn\n" ..
+                        AF_TAB .. "- {covenant}\n" ..
+                        AF_TAB .. AF_TAB .. "- {neutral}: Show achievements that all covenants can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {kyrian}: Show achievements only {kyrian_wc} can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {venthyr}: Show achievements only {venthyr_wc} can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {nightfae}: Show achievements only {nightfae_wc} can earn\n" ..
+                        AF_TAB .. AF_TAB .. "- {necrolord}: Show achievements only {necrolord_wc} can earn\n" ..
+                        AF_TAB .. "- {collapseseries}: Show only the last completed and first not completed achievement in a series\n\n" ..
                         "Sorting:\n" ..
                         AF_TAB .. "- {noSorting}: No sorting applied\n" ..
                         AF_TAB .. "- {sortByName}: Sort achievements based on their names\n" ..