+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [9.0.5.15.0] - 2021-03-23
+### Added
+- Warlords of Draenor Garrison (151 achievements)
+- Missing Legion Legion Class Hall (1 achievement)
+- Current Zone category (for now always visible and need to close/open the window or select the category again in order to refresh)
+- All Warlords of Draenor achievements should have zone data except for Pet Battles and the shared zone group
+- All Legion achievements should have zone data except for Pet Battles and the shared zone group
+- All Battle for Azeroth achievements should have zone data except for Pet Battles and the shared zone group
+- All Shadowlands achievements should have zone data except for Pet Battles
+- Filter to collapse a series of achievements so only the last completed and first not completed are visible
+
+### Changed
+- Battle for Azeroth Island Expeditions are now split per Island Expeditions
+- Battle for Azeroth War Effort are now split per War Effort
+
 ## [9.0.5.14.0] - 2021-03-10
 ### Added
 - Legion Pet Battles (with external links) (25 achievements)
@@ -181,6 +196,7 @@ All notable changes to this project will be documented in this file.
 ### Removed
 - The need for almost all localization
 
+[9.0.5.15.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3248934
 [9.0.5.14.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3234117
 [9.0.2.13.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3215749
 [9.0.2.12.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3205404